chore(box): move box voter to the box namespace
